Refactor the helper methods from DatabaseService to DatabaseHelpers
