+++ Server/Controllers/CustomizationController.cs
@@ -1,9 +1,9 @@
 ﻿using BSGHelperLibrary.ResponseModels;
 using Microsoft.AspNetCore.Mvc;
 using Paulov.Tarkov.WebServer.DOTNET.Middleware;
+using Paulov.TarkovServices.Helpers;
 using Paulov.TarkovServices.Providers.Interfaces;
 using Paulov.TarkovServices.Providers.SaveProviders;
-using Paulov.TarkovServices.Services;
 
 namespace Paulov.Tarkov.WebServer.DOTNET.Controllers
 {
@@ -19,7 +19,7 @@ namespace Paulov.Tarkov.WebServer.DOTNET.Controllers
         [HttpPost]
         public async void Customization(int? retry, bool? debug)
         {
-            DatabaseService.TryLoadCustomization(out var items);
+            DatabaseHelpers.TryLoadCustomization(out var items);
 
             await HttpBodyConverters.CompressIntoResponseBodyBSG(items, Request, Response);
         }
@@ -28,7 +28,7 @@ namespace Paulov.Tarkov.WebServer.DOTNET.Controllers
         [HttpPost]
         public IActionResult Storage(int? retry, bool? debug)
         {
-            DatabaseService.TryLoadDatabaseFile("database/templates/customisationStorage.json", out string file);
+            DatabaseHelpers.TryLoadDatabaseFile("database/templates/customisationStorage.json", out string file);
 
             return new BSGSuccessBodyResult(file);
         }
