+++ Server/Controllers/api/v1/ItemSearchController.cs
@@ -1,15 +1,10 @@
-﻿using System.Collections;
-using System.Collections.Concurrent;
-using System.Diagnostics;
-using BSGHelperLibrary.ResponseModels;
-using EFT.InventoryLogic;
+﻿using BSGHelperLibrary.ResponseModels;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.WebUtilities;
 using Microsoft.Extensions.Primitives;
 using Newtonsoft.Json.Linq;
-using Paulov.Tarkov.WebServer.DOTNET.Middleware;
-using Paulov.TarkovServices.Services;
-using Sirenix.Serialization;
+using Paulov.TarkovServices.Helpers;
+using System.Collections.Concurrent;
 
 namespace Paulov.Tarkov.WebServer.DOTNET.Controllers.api.v1
 {
@@ -24,31 +19,31 @@ namespace Paulov.Tarkov.WebServer.DOTNET.Controllers.api.v1
 
             int start = int.Parse(decodedForm["start"].First());
             int length = int.MaxValue;
-            if (decodedForm.TryGetValue("length", out  StringValues lengthValues))
+            if (decodedForm.TryGetValue("length", out StringValues lengthValues))
             {
                 length = int.Parse(lengthValues.First());
             }
-            
+
             //Localization pre-loading
             IEnumerable<KeyValuePair<string, JToken>> enumerableLocalizations =
-                DatabaseService.LoadDatabaseFileAsEnumerable("locales/global/en.json");
+                DatabaseHelpers.LoadDatabaseFileAsEnumerable("locales/global/en.json");
             Dictionary<string, string> localizationDictionary =
                 enumerableLocalizations.ToDictionary(localization =>
                     localization.Key, localization => localization.Value.ToString());
 
             //Price pre-loading
             IEnumerable<KeyValuePair<string, JToken>> enumerablePrices =
-                DatabaseService.LoadDatabaseFileAsEnumerable("templates/prices.json");
+                DatabaseHelpers.LoadDatabaseFileAsEnumerable("templates/prices.json");
             Dictionary<string, long> pricesDictionary =
                 enumerablePrices.ToDictionary(price =>
                     price.Key, price => long.Parse(price.Value.ToString()));
-            
+
             //Item pre-loading
             IEnumerable<KeyValuePair<string, JToken>> enumerableItems =
-                DatabaseService.LoadDatabaseFileAsEnumerable("templates/items.json");
+                DatabaseHelpers.LoadDatabaseFileAsEnumerable("templates/items.json");
             List<MinimalTemplateItem> items = enumerableItems.Select(x =>
                 new MinimalTemplateItem(x.Value, pricesDictionary.GetValueOrDefault(x.Key))).ToList();
-            
+
             //Highest price
             long highestPrice = items.Max(x => x.Price);
 
@@ -81,14 +76,14 @@ namespace Paulov.Tarkov.WebServer.DOTNET.Controllers.api.v1
                     const double basePvERarityMultiplier = 2.5;
 
                     Enum.TryParse(item.PvERarity, true, out PvERarity rarity);
-                    
+
                     priceRatio *= basePvERarityMultiplier + ((int)rarity * 0.25);
                     priceRatio = Math.Max(Math.Min(priceRatio, 100), 1);
                     priceRatio = Math.Ceiling(priceRatio);
                 }
 
                 if (rootResponseObject.Count > length) break;
-                
+
                 rootResponseObject.Add(new JObject
                 {
                     ["itemId"] = item.ItemID,
@@ -100,7 +95,7 @@ namespace Paulov.Tarkov.WebServer.DOTNET.Controllers.api.v1
                     ["priceRatio"] = priceRatio
                 });
             }
-            
+
             return new BSGSuccessBodyResult(rootResponseObject);
         }
 
@@ -111,17 +106,17 @@ namespace Paulov.Tarkov.WebServer.DOTNET.Controllers.api.v1
         {
             //Localization pre-loading
             IEnumerable<KeyValuePair<string, JToken>> enumerableLocalizations =
-                DatabaseService.LoadDatabaseFileAsEnumerable("locales/global/en.json");
+                DatabaseHelpers.LoadDatabaseFileAsEnumerable("locales/global/en.json");
             ConcurrentDictionary<string, string> localizationDictionary =
                 new(enumerableLocalizations.Select(localization =>
                     new KeyValuePair<string, string>(localization.Key, localization.Value.ToString())));
-            
-            
+
+
             //Item pre-loading
             IEnumerable<MinimalTemplateItem> templatesItemsMinimalEnumerable =
-                DatabaseService.LoadDatabaseFileAsEnumerable("templates/items.json")
+                DatabaseHelpers.LoadDatabaseFileAsEnumerable("templates/items.json")
                     .Select(x => new MinimalTemplateItem(x.Value, 0));
-            
+
             const string ammoParentId = "5485a8684bdc2da71d8b4567";
             string[] ammoIdsToIgnore = ["5996f6d686f77467977ba6cc", "5d2f2ab648f03550091993ca", "5cde8864d7f00c0010373be1"];
 
@@ -131,7 +126,7 @@ namespace Paulov.Tarkov.WebServer.DOTNET.Controllers.api.v1
                 if (!string.Equals(item.ParentID, ammoParentId)) return ValueTask.CompletedTask;
                 if (ammoIdsToIgnore.Contains(item.ItemID)) return ValueTask.CompletedTask;
                 if (!string.Equals(item.Props.AmmoType, "bullet")) return ValueTask.CompletedTask;
-                
+
                 //Localized name
                 string localizedItemName = item.ItemID;
                 if (!localizationDictionary.TryGetValue($"{item.ItemID} Name", out localizedItemName))
@@ -140,7 +135,7 @@ namespace Paulov.Tarkov.WebServer.DOTNET.Controllers.api.v1
                 }
 
                 int rating = 0; // Math.ceil((item.PaulovRating / highestRating) * 100);
-                
+
                 rootResponseObject.Add(new JObject
                 {
                     ["itemId"] = item.ItemID,
